Adding deadband, changing tuning file header
There is now a deadband in the PID routine for any temperature greater than 0.25C above the setpoint, output is set to go to zero in that case, since the mussels have no active cooling capability. The tuning output file also gains a spot in the 1st header line to record the target temperature increase.